Fixes #4831: Change id reports type to bigint in psql, and Long in Scala

+++ rudder-core/src/main/resources/reportsSchema.sql
@@ -78,7 +78,7 @@ create index expectedReportsNodes_versionId on expectedReportsNodes (nodeJoinKey
 create sequence serial START 101;
 
 CREATE TABLE RudderSysEvents (
-id integer PRIMARY KEY default nextval('serial'),
+id bigint PRIMARY KEY default nextval('serial'),
 executionDate timestamp with time zone NOT NULL, 
 nodeId text NOT NULL CHECK (nodeId <> ''),
 directiveId text NOT NULL CHECK (directiveId <> ''),
@@ -102,7 +102,7 @@ CREATE INDEX ruleId_idx on RudderSysEvents (ruleId);
 
 
 CREATE TABLE ArchivedRudderSysEvents (
-id integer PRIMARY KEY,
+id bigint PRIMARY KEY,
 executionDate timestamp with time zone NOT NULL, 
 nodeId text NOT NULL CHECK (nodeId <> ''),
 directiveId text NOT NULL CHECK (directiveId <> ''),
@@ -278,7 +278,7 @@ CREATE TABLE Workflow(
 
 create table StatusUpdate (
 key text PRIMARY KEY,
-lastId integer NOT NULL,
+lastId bigint NOT NULL,
 date timestamp with time zone NOT NULL
 );
